pyproject.toml
[tool.poetry]
name = "get-papers"
version = "0.1.0"
description = "Fetch PubMed papers with pharmaceutical affiliations"
authors = ["Prashant Kumar <your_email@example.com>"]
packages = [{ include = "papers_fetcher" }]

[tool.poetry.dependencies]
python = "^3.10"
biopython = "^1.83"
typer = "^0.12.3"
rich = "^13.7.1"

[tool.poetry.scripts]
get-papers-list = "cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
